milestone2/backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
tavily-python
fpdf2
python-dotenv
beautifulsoup4
pydantic
requests
markdown
